--- src/main/java/com/wwi21sebgroup5/cinema/config/UserController.java
@@ -1,22 +1,0 @@
-package com.wwi21sebgroup5.cinema.config;
-
-import com.wwi21sebgroup5.cinema.services.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(path = "/v1/user")
-public class UserController {
-
-    @Autowired
-    UserService userService;
-
-    @PostMapping(path = "/register")
-    public ResponseEntity<Object> register() {
-        return userService.register();
-    }
-
-}
