+++ src/main/java/com/wwi21sebgroup5/cinema/controller/RoleController.java
@@ -1,15 +1,15 @@
-package com.wwi21sebgroup5.cinema.config;
+package com.wwi21sebgroup5.cinema.controller;
 
-import com.wwi21sebgroup5.cinema.services.CityService;
+import com.wwi21sebgroup5.cinema.services.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping(path = "/v1/city")
-public class CityController {
+@RequestMapping(path = "/v1/role")
+public class RoleController {
 
     @Autowired
-    CityService cityService;
+    RoleService roleService;
 
 }
