+++ src/main/java/com/wwi21sebgroup5/cinema/controller/UserController.java
@@ -1,0 +1,41 @@
+package com.wwi21sebgroup5.cinema.controller;
+
+import com.wwi21sebgroup5.cinema.entities.User;
+import com.wwi21sebgroup5.cinema.services.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@RestController
+@RequestMapping(path = "/v1/user")
+public class UserController {
+
+    @Autowired
+    UserService userService;
+
+    @PostMapping(path = "/register")
+    public ResponseEntity<Object> register() {
+        return userService.register();
+    }
+
+    @PostMapping(path = "/login")
+    public ResponseEntity<Object> login() {
+        return userService.login();
+    }
+
+    @GetMapping(path = "/{id}")
+    public ResponseEntity<User> getUserById(@PathVariable UUID id) {
+        Optional<User> user = userService.getUserById(id);
+
+        if (user.isPresent()) {
+            return new ResponseEntity<>(user.get(), HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+}
