+++ src/main/java/com/wwi21sebgroup5/cinema/services/UserService.java
@@ -1,9 +1,13 @@
 package com.wwi21sebgroup5.cinema.services;
 
+import com.wwi21sebgroup5.cinema.entities.User;
 import com.wwi21sebgroup5.cinema.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+import java.util.UUID;
 
 @Service
 public class UserService {
@@ -11,7 +15,15 @@ public class UserService {
     @Autowired
     UserRepository userRepository;
 
+    public Optional<User> getUserById(UUID id) {
+        return userRepository.findById(id);
+    }
+
     public ResponseEntity<Object> register() {
+        return null;
+    }
+
+    public ResponseEntity<Object> login() {
         return null;
     }
 }
