+++ src/main/resources/application-dev.properties
@@ -1,5 +1,4 @@
 server.port=8082
-spring.h2.console.enabled=true
 
 spring.datasource.url=jdbc:mysql://localhost:3306/cinema
 spring.datasource.username=root
